api/requirements.txt
langchain
langchain-openai
langchain-community
openai
fastapi
uvicorn
requests
python-dotenv